[CHORE] Include Reader application and reorganize repository
With this commit, the reader application that was formerly at
github.com/manifoldScholar/reader is now brought into this project.
No attempt has been made to recover the commit history of the reader,
as it was by and large developed as a proof of concept.

Moreover, this commit adds placeholder modules for the Manifold backend,
frontend, and skin. With Manifold, we hope to avoid creating a large,
Monolithic code base, which is why the project is broken into discrete
modules with some shared functionality. However, we are choosing to
kepe thes modules in a single repository, rather than using Git subtrees,
Git submodules, or separate projects. Our hope is that the single
repository will reduce friction for new developers, allow us to centralize
issue tracking, and to coordinate revisions across modules.