[CHORE] Add a script module; add NPM autolinking script
This commit adds a new module to the project, which will be used to
store scripts and other glue scripts related to setting up a Manifold
development environment.

The first script being added is the setup script, which should be used
to perform any tasks that are needed as part of an initial setup. At
some future point, we may want to use gulp or grunt as a task runner,
but my hunch is that we may not need what gulp brings us, and I'd
prefer not to introduce another build mechanism when we already have
Webpack for the node modules.

The second script this commit adds is link.rb. This script will look
through each project module, determine if it is a node module,
create a global NPM link for it, and then include it in each
node_module as a symlink. See https://docs.npmjs.com/cli/link for more
on how this works.

For now, all NPM packages on the project will be cross linked to one
another, which allows easy includes of exported resources across
packages. Requiring files across node modules should be used with
caution, of course.

The reasoning behind this approach is twofold. First, we'd like to be
able to share style assets between projects, but we don't want to
update node modules after every change. Second, we're anticipating that
there will be shared utility and library modules, which will need linking.